+++ package.json
@@ -16,7 +16,6 @@
     "@mui/material": "^5.15.16",
     "@preset-sdk/embedded": "^0.1.10",
     "axios": "^1.6.8",
-    "emoji-mart": "^5.6.0",
     "emoji-picker-react": "^4.9.3",
     "next": "^14.2.3",
     "react": "^18.3.1",
