+++ src/app/globals.css
@@ -79,10 +79,10 @@
   margin: 0;
 }
 
-html,
-body {
-  max-width: 100vw;
-  overflow-x: hidden;
+html, body{
+  overflow-x: hidden !important;
+  overflow-y: auto !important;
+  scroll-behavior: smooth;
 }
 
 a {
@@ -93,6 +93,8 @@ a {
 iframe {
   width: 100%;
   height: 100%;
+  border: none;
+  overflow: hidden;
 }
 /*@media (prefers-color-scheme: dark) {*/
 /*  html {*/
