+++ src/app/page.js
@@ -1,123 +1,85 @@
-'use client';
+'use client'
 
 import React, {useEffect, useState} from 'react';
-import {
-  TextField,
-  Grid,
-  Link,
-  Card,
-  OutlinedInput,
-  InputLabel,
-  InputAdornment,
-  IconButton, FormControl, Box
-} from '@mui/material';
-import {Visibility, VisibilityOff} from "@mui/icons-material";
-import useAuthStore from "@/app/store/auth";
-import {useSnackbar} from "@/app/context/SnackbarContext";
-import cache from "@/app/services/cache";
-import {LoadingButton} from "@mui/lab";
-import { useRouter } from 'next/navigation'
+import {Box, Grid, Paper, Typography, List, ListItem, ListItemText, Button, Card} from '@mui/material';
+import useNotificationStore from "@/app/store/notification";
+import useDebtsStore from "@/app/store/debts";
 
-export default function Home() {
+const Dashboard = () => {
+    const [notifications, setNotifications] = useState([]);
 
-  const { showSnackbar } = useSnackbar();
-  const { login } = useAuthStore();
-  const router = useRouter();
+    const { getNotifications } = useNotificationStore();
+    const { getMyDebts, debts } = useDebtsStore();
 
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [showPassword, setShowPassword] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
+    useEffect(() => {
+        async function load() {
+            setNotifications(await getNotifications());
+            await getMyDebts();
+        }
+        load();
+    }, []);
 
-  useEffect(() => {
-    if (cache.get('token')) {
-      router.replace('/groups');
-    }
-  }, [router]);
+    return (
+        <Box sx={{ flexGrow: 1, padding: 2, minHeight: '100vh' }}>
+            <Grid container spacing={2}>
+                <Grid item xs={12} sm={6} md={8}>
+                    <Card sx={{ padding: 2, marginBottom: 2 }}>
+                        <Typography variant="h6">Latest activities</Typography>
+                        <List>
+                            {notifications.slice().reverse().slice(0,5).map((notification) => (
+                                <ListItem>
+                                    <ListItemText
+                                        primary={new Date(notification.createdAt).toLocaleString()}
+                                        secondary={notification.message}
+                                    />
+                                </ListItem>
+                            ))}
+                        </List>
+                    </Card>
 
-  const handleEmailChange = (event) => {
-    setEmail(event.target.value);
-  };
+                    <Card>
+                        <Typography variant="h6">Favorites</Typography>
+                        <List>
+                            <ListItem>
+                                <Button variant="contained" fullWidth>
+                                    Group a
+                                </Button>
+                            </ListItem>
+                            <ListItem>
+                                <Button variant="contained" fullWidth>
+                                    Group b
+                                </Button>
+                            </ListItem>
+                            <ListItem>
+                                <Button variant="contained" fullWidth>
+                                    Group c
+                                </Button>
+                            </ListItem>
+                        </List>
+                    </Card>
+                </Grid>
+                <Grid item xs={12} sm={6} md={4}>
+                    <Card sx={{ padding: 2, marginBottom: 2 }}>
+                        <Typography variant="h6">Debts</Typography>
+                        <List>
+                            {debts.slice().reverse().slice(0,5).map((debt) => (
+                                <ListItem>
+                                    <ListItemText
+                                        primary={debt.userToId}
+                                        secondary={debt.amount - debt.amountPaid}
+                                    />
+                                </ListItem>
+                            ))}
+                        </List>
+                    </Card>
 
-  const handlePasswordChange = (event) => {
-    setPassword(event.target.value);
-  };
+                    <Paper sx={{ padding: 2, marginBottom: 2 }}>
+                        <Typography variant="h6">Some information and quick access to reports</Typography>
+                    </Paper>
+                </Grid>
+            </Grid>
+        </Box>
+    );
+};
 
-  const handleClickShowPassword = () => {
-    setShowPassword((show) => !show);
-  };
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    setIsLoading(true)
-
-    try{
-      await login({
-        email,
-        password
-      });
-
-      await router.push('/groups');
-    } catch (error) {
-      showSnackbar(error.response.data.error, 'error');
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <Grid container alignItems="center" justifyContent="center" style={{ height: '100vh' }}>
-      <Grid item xs={12} sm={6} md={4}>
-        <Card variant="outlined" sx={{ p: 4}}>
-          <h2>Login</h2>
-          <form onSubmit={handleSubmit}>
-            <TextField
-              label="Email"
-              type="email"
-              fullWidth
-              margin="normal"
-              variant="outlined"
-              value={email}
-              onChange={handleEmailChange}
-              required
-            />
-            <FormControl fullWidth variant="outlined">
-              <InputLabel htmlFor="outlined-adornment-password">Contraseña</InputLabel>
-              <OutlinedInput
-                id="outlined-adornment-password"
-                type={showPassword ? 'text' : 'password'}
-                value={password}
-                onChange={handlePasswordChange}
-                required
-                endAdornment={
-                  <InputAdornment position="end">
-                    <IconButton
-                      aria-label="toggle password visibility"
-                      onClick={handleClickShowPassword}
-                      edge="end"
-                    >
-                      {showPassword ? <VisibilityOff /> : <Visibility />}
-                    </IconButton>
-                  </InputAdornment>
-                }
-                label="Contraseña"
-              />
-            </FormControl>
-            <Box display="flex" justifyContent="center" marginTop={2} marginBottom={2}>
-              <LoadingButton
-                size="large"
-                loading={isLoading}
-                variant="outlined"
-                color="secondary"
-                type="submit"
-              >
-                <span>Iniciar sesión</span>
-              </LoadingButton>
-            </Box>
-          </form>
-          <hr />
-          <p style={{ marginTop: 20 }}>¿Aún no tienes una cuenta? <Link color = "secondary" href="/register">Registrarse</Link></p>
-        </Card>
-      </Grid>
-    </Grid>
-  );
-}
+export default Dashboard;
